\name{chooseC}
\alias{chooseC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Binomial coefficient
}
\description{
It calculates the binomial coefficient.
}
\usage{
chooseC(n, k)
}
\arguments{
  \item{n}{
set size
}
\item{k}{
number of unordered outcomes
}
}
\details{
implemented in C++
}
\value{
numeric
}
\examples{
chooseC(n=6, k=2)
}