\name{svisits-package}
\alias{svisits-package}
\alias{svisits}
\docType{package}
\title{
svisits: finding HIV transmission and serosorting pairs using shared clinic visit dates}
\description{
  This package helps to find some stable HIV-infected partnerships in cohort studies based on the assumption that some patients frequently attend the clinical visits together. These pairs are useful for biological and epidemiological research.
}
\author{
Alex Marzel, Teja Turk, Roger Kouyos

Maintainer: Alex Marzel <Alex.Marzel@usz.ch>
}
\keyword{package}
\examples{
  \dontrun{
       
  }
}
