\name{chooseC}
\alias{chooseC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Binomial coefficient
}
\description{
Calculates the Binomial coefficient
}
\usage{
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
size
}
\item{k}{
number of unordered outcomes
}
}
\details{
}
\value{
number
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
chooseC(n=6,k=2)
# 15
}