\name{svisits-package}
\alias{svisits-package}
\alias{svisits}
\docType{package}
\title{
svisits: finding HIV transmission and serosorting pairs using shared clinic visit dates}
\description{
  This package helps to find some stable HIV-infected partnerships in cohort studies basing on the assumption that some patients frequently attend the clinical visits together. These pairs are useful for biological and epidemiological research.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Alex Marzel, Teja Turk, Roger Kouyos

Maintainer: Alex Marzel <Alex.Marzel@usz.ch>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
       
  }
}
